rename files paths
